Move manipulation os env vars to prefix data
