+++ conda/base/constants.py
@@ -228,8 +228,8 @@ class SatSolverChoice(Enum):
 PACKAGE_CACHE_MAGIC_FILE = 'urls.txt'
 PREFIX_MAGIC_FILE = join('conda-meta', 'history')
 
-PREFIX_SATE_FILE = join('conda-meta', 'state.json')
-PACKAGE_ENV_VARS_DIR = join('etc', 'conda', 'env_vars')
+PREFIX_SATE_FILE = join('conda-meta', 'state')
+PACKAGE_ENV_VARS_DIR = join('etc', 'conda', 'env_vars.d')
 
 
 # TODO: should be frozendict(), but I don't want to import frozendict from auxlib here.
