+++ tests/core/test_prefix_data.py
@@ -4,14 +4,30 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from contextlib import contextmanager
-import os
-from os.path import isdir
+from os.path import isdir, join, lexists
+from tempfile import gettempdir
+from unittest import TestCase
+from uuid import uuid4
 
 from conda.common.compat import on_win, odict
 from conda.core.prefix_data import PrefixData, get_conda_anchor_files_and_records
 from test_data.env_metadata import (
     PATH_TEST_ENV_1, PATH_TEST_ENV_2, PATH_TEST_ENV_3, PATH_TEST_ENV_4,
 )
+from conda.base.constants import PREFIX_SATE_FILE
+from conda.gateways.disk import mkdir_p
+from conda.gateways.disk.delete import rm_rf
+
+
+ENV_VARS_FILE = '''
+{
+  "version": 1,
+  "env_vars": {
+    "ENV_ONE": "one",
+    "ENV_TWO": "you",
+    "ENV_THREE": "me"
+  }
+}'''
 
 
 def _print_output(*args):
@@ -164,3 +180,50 @@ def test_get_conda_anchor_files_and_records():
 
     _print_output(output, expected_output)
     assert output == expected_output
+
+class PrefixDatarUnitTests(TestCase):
+
+    def setUp(self):
+        tempdirdir = gettempdir()
+        dirname = str(uuid4())[:8]
+        self.prefix = join(tempdirdir, dirname)
+        mkdir_p(self.prefix)
+        assert isdir(self.prefix)
+        mkdir_p(join(self.prefix, 'conda-meta'))
+        activate_env_vars = join(self.prefix, PREFIX_SATE_FILE)
+        with open(activate_env_vars, 'w') as f:
+            f.write(ENV_VARS_FILE)
+        self.pd = PrefixData(self.prefix)
+
+    def tearDown(self):
+        rm_rf(self.prefix)
+        assert not lexists(self.prefix)
+
+    def test_get_environment_env_vars(self):
+        ex_env_vars = {
+            "ENV_ONE": "one",
+            "ENV_TWO": "you",
+            "ENV_THREE": "me"
+        }
+        env_vars = self.pd.get_environment_env_vars()
+        assert ex_env_vars == env_vars
+
+    def test_set_unset_environment_env_vars(self):
+        env_vars_one = {
+            "ENV_ONE": "one",
+            "ENV_TWO": "you",
+            "ENV_THREE": "me",
+        }
+        env_vars_add = {
+            "ENV_ONE": "one",
+            "ENV_TWO": "you",
+            "ENV_THREE": "me",
+            "WOAH": "dude"
+        }
+        self.pd.set_environment_env_vars({"WOAH":"dude"})
+        env_vars = self.pd.get_environment_env_vars()
+        assert env_vars_add == env_vars
+
+        self.pd.unset_environment_env_vars(['WOAH'])
+        env_vars = self.pd.get_environment_env_vars()
+        assert env_vars_one == env_vars
